+++ plugins/scheduler-docker-local/.gitignore
@@ -1,0 +1,3 @@
+/triggers/*
+/triggers
+/cron-*
