+++ plugins/scheduler-docker-local/Makefile
@@ -1,0 +1,5 @@
+TRIGGERS = triggers/cron-write
+BUILD = triggers
+PLUGIN_NAME = scheduler-docker-local
+
+include ../../common.mk
