+++ plugins/scheduler-docker-local/post-delete
@@ -14,6 +14,8 @@ trigger-scheduler-docker-local-post-delete() {
 
   # remove all application containers
   "$DOCKER_BIN" container ls --filter "label=com.dokku.app-name=${APP}" -q | xargs -n1 -I {} "$DOCKER_BIN" container rm --force {} &>/dev/null || true
+
+  plugn trigger cron-write
 }
 
 trigger-scheduler-docker-local-post-delete "$@"
