+++ plugins/scheduler-docker-local/triggers.go
@@ -1,0 +1,6 @@
+package schedulerdockerlocal
+
+// TriggerCronWrite force updates the cron file for all apps
+func TriggerCronWrite() error {
+	return writeCronEntries()
+}
